requirement.txt
langchain
chromadb
tiktoken
langchain-community
ctransformers==0.2.5 
sentence-transformers
pypdf
pandas
numpy
python-dotenv
langchain-huggingface
PyMuPDF
fastembed
deepgram-sdk
SpeechRecognition
asyncio==3.4.3
PyAudio
langchain-groq
ffmpeg
ffmpeg-python
streamlit